Merge dev to master

+++ Store.Application/UseCases/Categories/DeleteCategoryUseCase.cs
@@ -7,9 +7,13 @@ namespace Store.Application.UseCases.Categories
         private readonly ICategoryRepository CategoryRepository = categoryRepository;
 
 
-        public async Task Execute(int id)
+        public async Task<int> Execute(int id)
         {
+            if (await CategoryRepository.IsUsed(id))
+                throw new InvalidOperationException("Other Category or Product still use this category, cannot delete");
+
             await CategoryRepository.DeleteAsync(id);
+            return id;
         }
     }
 }
